bsps: Fix integer types in bsp_fdt_copy()
Update #3082.

+++ c/src/lib/libbsp/shared/src/bsp-fdt.c
@@ -47,9 +47,9 @@ void bsp_fdt_copy(const void *src)
 #endif
 
   if (s != d) {
-    uint32_t m = MIN(sizeof(bsp_fdt_blob), fdt_totalsize(src));
-    uint32_t n = (m + sizeof(*d) - 1) / sizeof(*d);
-    uint32_t i;
+    size_t m = MIN(sizeof(bsp_fdt_blob), fdt_totalsize(src));
+    size_t n = (m + sizeof(*d) - 1) / sizeof(*d);
+    size_t i;
 
     for (i = 0; i < n; ++i) {
       d[i] = s[i];
